Renaming the autolog file from pytorch-lightning-autolog to pytorch-autolog
